requirements.txt
os
tensorflow
Pillow
numpy
sklearn
pandas
matplotlib
seaborn
streamlit
opencv-python-headless
